Create Kaplan-Meier for Survival Analysis
This pertains to an activity within a Master's program in Applied Statistics.

Activity on Theme 4:

The following table presents data from a clinical trial on the efficacy of azidothymidine (AZT) treatment for patients with acquired immunodeficiency syndrome (AIDS).

The trial commenced in February 1987 and concluded in February 1989. A total of 32 patients in an advanced stage of the disease were included. Sixteen patients were randomly assigned to the AZT-treated group, while another sixteen patients were allocated to the placebo-receiving group.

Requested Tasks:
a) Obtain, for each patient, their participation time (in months), indicating whether it's a complete or censored data point.
b) Construct the life table considering semesters as time intervals.
c) Estimate the survival function using the Kaplan-Meier estimator. Obtain an estimation of errors and interpret the results.
d) Obtain survival curves for the data classified by treatment groups and compare them. Explain the results.