Update and rename KM_Survival.R to Kaplan-Meier for Survival Analysis/KM_Survival.R
This pertains to an activity within a Master's program in Applied Statistics.

Activity on Theme 4:

The following dataset presents information from a clinical trial studying the efficacy of azidothymidine (AZT) treatment in patients with acquired immunodeficiency syndrome (AIDS). The trial began in February 1987 and concluded in February 1989. A total of 32 patients in advanced stages of the disease were included. Sixteen patients were randomly assigned to the group receiving AZT, and the remaining sixteen were assigned to the placebo group.

You are required to:

a) Obtain for each patient their participation time (in months), indicating whether the data is complete or censored.
b) Build a life table using six-month intervals.
c) Estimate the survival function using the Kaplan-Meier estimator. Include error estimation and interpret the results.
d) Obtain and compare survival curves for the data classified by treatment group. Explain the findings.