Update and rename COX_Survival.R to Cox for Survival Analysis/COX_Survival.R
This pertains to an activity within a Master's program in Applied Statistics.

Activity on Theme 5:

The random variable under consideration is the duration of unemployment. It is known that certain factors decisively influence the time an individual remains unemployed. A sample of 40 individuals registered at the employment office of a specific locality is considered, and the time elapsed until they either cease to be unemployed by finding a job (censorship=1) or cease to appear on the list of unemployed due to some other reason (censorship=0) is recorded. Additionally, information regarding their age, gender, and professional sector is recorded for each individual. The data is presented in the attached table.

Fit a proportional hazards model with a constant baseline hazard function and discuss the results.